# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/review": {
            "post": {
                "description": "Tạo review từ user đến user khác. Review phải được đính với id session đấu già mà user đó thắng. Mỗi id session chỉ có duy nhất 1 review. Một user không thể review chính bản thân mình.",
                "tags": [
                    "Review Controller"
                ],
                "summary": "Tạo review mới",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Information to be provided",
                        "name": "NewSessionInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NewReview"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "401": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/review/:sessionid": {
            "delete": {
                "tags": [
                    "Review Controller"
                ],
                "summary": "Xóa review của user dựa theo session id (administrator only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID session",
                        "name": "sessionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "400": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "401": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/review/:userid": {
            "get": {
                "description": "Lấy review bằng userid dưới dạng path, trả về form JSON.",
                "tags": [
                    "Review Controller"
                ],
                "summary": "Lấy review bằng userid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UserReview"
                        }
                    },
                    "500": {
                        "description": "Error message",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.NewReview": {
            "type": "object",
            "required": [
                "reviewcontent",
                "reviewscore",
                "sessionid",
                "usertargetid"
            ],
            "properties": {
                "reviewcontent": {
                    "type": "string"
                },
                "reviewscore": {
                    "type": "integer"
                },
                "sessionid": {
                    "type": "integer"
                },
                "usertargetid": {
                    "type": "string"
                }
            }
        },
...
        "model.UserReview": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "sessionid": {
                    "type": "integer"
                },
                "targetid": {
                    "type": "string"
                },
                "writerid": {
                    "type": "string"
                }
            }
        }
    }
}